Update enrollment status when academic record is inserted

+++ enrollment-service/src/main/java/com/project/enrollmentservice/controller/EnrollmentController.java
@@ -51,7 +51,7 @@ public class EnrollmentController {
 
     @GetMapping("/students/{studentId}")
     public ResponseEntity<List<EnrollmentResponseDTO>> getEnrollmentsByStudent(
-            @PathVariable int studentId,
+            @PathVariable Long studentId,
             @RequestParam(required = false) List<String> status) {
 
         if (status == null || status.isEmpty()) {
